+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Line wrap the file at 100 chars.                                              Th
 - Do not show dialog when rendering the map fails due to disabled GPU acceleration.
 - Require the post-quantum X25519MLKEM768 key exchange for TLS connections to the Mullvad API.
 - Disable TLS session tickets to reduce the ability to track clients over time.
+- Remove old log files, such as OpenVPN and wireguard-go logs.
 
 #### Linux
 - Remove dependency on `iproute2` when using GotaTun with IPv6.